consumer/requirement.txt
sqlalchemy
psycopg2-binary
kafka-python
pydantic
python-dotenv